requirements.txt
aiogram==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.8.4
aiosignal==1.4.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
propcache==0.3.2
python-dotenv==1.2.1
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
typing_extensions==4.14.1
urllib3==2.0.4
yarl==1.20.1
